Used the new properties widget in the Custom Types Editor
* Moved VariantMapProperty to its own file, because it is now also used
  by the PropertyTypesEditor.

* Introduced VariantEditorView to share a bit of code.
